CLUSTERING | BFR ALGORITHM | PYSPARK
*Implement the Bradley-Fayyad-Reina (BFR) algorithm. The goal is to let you
familiar with the process of clustering in general and various distance measurements. The dataset you
are going to use is the synthetic dataset.*

Since the BFR algorithm strongly assumes that the clusters are normally distributed with independent dimensions, we generated synthetic datasets by initializing some random centroids and creating some data points with the centroids and some standard deviations to form the clusters. We also added some other data points as outliers in the dataset to evaluate the algorithm. Data points that are outliers belong to clusters that are named or indexed as “-1”. Figure 1 shows an example of a part of the dataset. The first column is the data point index. The second column is the name index of the cluster that the data point belongs to. The rest columns represent the feature dimensions of the data point.